Fixed a bug in the Browse UI.
When the Browse page UI was engineered, the pagination component worked
by changing the URL when a different link was selected.

This had the side-effect of the page redirecting to the first page
whenever a link was changed because the Browse page detects changes to
the form elements to refresh inner xhr data.

As a workaround, to avoid redirects, there was a line in the form
watcher set up to ignore the page.

While this workaround worked, it presented challenges in adding local
storage because the form object was saved dynamically and overwriting
the form with local storage was redirecting the page back to the first
page of the pagination set.

This fix will properly use the links to send an event that tells the
form to change its page property, instead of hardcoding the href. Since
the page navigation now works by changing the form.page property, the
line that ignores paginatination from the form refresh can be removed.